Board コンポーネントを分離
